Fix command line args object not being the correct one
In beeref/selection.py and beeref/view.py the commandline arg parse
object would be created on import.

Previously this would lead to the arg parse object being out of date
when checking it.

Now we don't recreate the arg parse helper object and instead make sure
that it is static.